# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organisations/{organisationId}/workflows/definitions": {
      "get": {
        "operationId": "Workflows_definitions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Definitions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDefinitionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:read"
        ],
        "summary": "Registered workflow definitions",
        "description": "**Requires permission:** `workflow:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/workflows": {
      "post": {
        "operationId": "Workflows_start",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInstance"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:manage"
        ],
        "summary": "Start an API-startable workflow",
        "description": "**Requires permission:** `workflow:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartWorkflowRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "Workflows_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "description": "Page size"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "`nextCursor` from the previous page"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:read"
        ],
        "summary": "List workflow instances",
        "description": "**Requires permission:** `workflow:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/workflows/{workflowId}": {
      "get": {
        "operationId": "Workflows_get",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:read"
        ],
        "summary": "Workflow with history and the actions the caller may take",
        "description": "**Requires permission:** `workflow:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/workflows/{workflowId}/transitions": {
      "post": {
        "operationId": "Workflows_transition",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transitioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInstance"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:read"
        ],
        "summary": "Perform a transition (permission, segregation of duties and optimistic concurrency enforced)",
        "description": "**Requires permission:** `workflow:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWorkflowRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/workflows/{workflowId}/reassign": {
      "post": {
        "operationId": "Workflows_reassign",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reassigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInstance"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Workflows"
        ],
        "x-required-permissions": [
          "workflow:manage"
        ],
        "summary": "Reassign (or unassign) a workflow; recorded in the history",
        "description": "**Requires permission:** `workflow:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignWorkflowRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "UK Accounting Platform API",
    "description": "Practice-first accounting, tax and compliance platform. **V0 foundation:** identity, tenancy, RBAC, documents, jobs, audit, outbox events, workflows, tasks, notifications, integration and AI abstractions.\n\nAll tenant data lives under `/organisations/{organisationId}` and is protected by authentication, RBAC, company scope and PostgreSQL row-level security. Errors use RFC 9457 `application/problem+json`.",
    "version": "1.0.0",
    "contact": {},
    "x-api-versioning": "URI versioning (/api/v1). Additive changes only within a version; breaking changes ship as /api/v2."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Opaque session token from `POST /auth/login/bearer`"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uk_session",
        "description": "httpOnly SameSite=Strict cookie set by `POST /auth/login`. Unsafe methods must send an allowed `Origin`."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:uk-platform:error:validation_error"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "description": "Echoes `x-request-id`"
          }
        }
      },
...
      "WorkflowInstance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "definitionVersion": {
            "type": "integer"
          },
          "companyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "attempt": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          },
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "startedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "definitionVersion",
          "companyId",
          "assigneeUserId",
          "attempt",
          "state",
          "subjectType",
          "subjectId",
          "version",
          "startedByUserId",
          "completedAt"
        ],
        "additionalProperties": false
      },
      "WorkflowDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "definitionVersion": {
            "type": "integer"
          },
          "companyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "attempt": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          },
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "startedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transitions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fromState": {
                  "type": "string",
                  "nullable": true
                },
                "toState": {
                  "type": "string"
                },
                "action": {
                  "type": "string"
                },
                "actorUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "comment": {
                  "type": "string",
                  "nullable": true
                },
                "attempt": {
                  "type": "integer"
                },
                "evidenceDocumentIds": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "occurredAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "fromState",
                "toState",
                "action",
                "actorUserId",
                "comment",
                "attempt",
                "evidenceDocumentIds",
                "occurredAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "type",
          "definitionVersion",
          "companyId",
          "assigneeUserId",
          "attempt",
          "state",
          "subjectType",
          "subjectId",
          "version",
          "startedByUserId",
          "completedAt",
          "availableActions",
          "transitions"
        ],
        "additionalProperties": false
      },
...
      "WorkflowPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": {
                  "type": "string"
                },
                "definitionVersion": {
                  "type": "integer"
                },
                "companyId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "assigneeUserId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "attempt": {
                  "type": "integer"
                },
                "state": {
                  "type": "string"
                },
                "subjectType": {
                  "type": "string"
                },
                "subjectId": {
                  "type": "string"
                },
                "version": {
                  "type": "integer"
                },
                "startedByUserId": {
                  "type": "string",
                  "format": "uuid"
                },
                "completedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "type",
                "definitionVersion",
                "companyId",
                "assigneeUserId",
                "attempt",
                "state",
                "subjectType",
                "subjectId",
                "version",
                "startedByUserId",
                "completedAt"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "WorkflowDefinitionList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "version": {
                  "type": "integer"
                },
                "initialState": {
                  "type": "string"
                },
                "terminalStates": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "apiStartable": {
                  "type": "boolean"
                },
                "transitions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string"
                      },
                      "from": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "to": {
                        "type": "string"
                      },
                      "permission": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "action",
                      "from",
                      "to",
                      "permission"
                    ],
                    "additionalProperties": true
                  }
                }
              },
              "required": [
                "type",
                "version",
                "initialState",
                "terminalStates",
                "apiStartable",
                "transitions"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "StartWorkflowRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 60
          },
          "subjectType": {
            "type": "string",
            "maxLength": 60
          },
          "subjectId": {
            "type": "string",
            "maxLength": 100
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "context": {
            "type": "object",
            "additionalProperties": {},
            "default": {}
          }
        },
        "required": [
          "type",
          "subjectType",
          "subjectId"
        ],
        "additionalProperties": false
      },
      "TransitionWorkflowRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 60
          },
          "comment": {
            "type": "string",
            "maxLength": 2000
          },
          "expectedVersion": {
            "type": "integer",
            "exclusiveMinimum": true,
            "minimum": 0
          },
          "evidenceDocumentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 50
          }
        },
        "required": [
          "action"
        ],
        "additionalProperties": false
      },
      "ReassignWorkflowRequest": {
        "type": "object",
        "properties": {
          "assigneeUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "assigneeUserId"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "400": {
        "description": "Malformed request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "401": {
        "description": "Not authenticated or session expired",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "403": {
        "description": "Authenticated but not permitted (RBAC, segregation of duties, CSRF origin)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "404": {
        "description": "Not found — also returned for resources in other organisations (existence is never revealed)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "409": {
        "description": "Conflict with current state",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "422": {
        "description": "Validation failed (`errors[]` lists fields)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}